# Various optimizations for the DataFlowAnalyzer.