… libsolutil/CMakeLists.txt
… libyul/optimiser/Rematerialiser.cpp
… libyul/optimiser/StackCompressor.cpp