Part of #10405

I'm not necessarily saying that any of this is a particularly good idea :-)...
But with this I get a speedup of almost 15% on a RelWithDebInfo build:

develop:
```
$ time { for i in $(seq 10); do build/test/tools/isoltest -t sem*/via* --optimize &> /dev/null ; done;}

real    0m47,327s
user    0m46,901s
sys     0m0,294s
```

this branch:
```
$ time { for i in $(seq 10); do build/test/tools/isoltest -t sem*/via* --optimize &> /dev/null ; done;}

real    0m40,649s
user    0m40,286s
sys     0m0,259s
```

But I'm not sure how much further one can get with micro-optimizations like this.