@ekpyron 2020-12-01
The most expensive remaining part when I profile this branch is ``DataFlowAnalyzer::handleAssignment``, so if there is much further to be done that's where to look.