@ekpyron 2020-12-10
> What is the status of merging this?

We'll probably merge some version at some point, but some comments etc. still need to be addressed. My plan was to take care of this one tomorrow.