@ekpyron 2020-12-11
Should still be squashed - and it would be nice, if someone tried to confirm that I'm right in that not clearing back-references when popping scopes is fine, since we check for variables being in scope on the user end (see https://github.com/ethereum/solidity/pull/10467#discussion_r533790328)