@ekpyron 2021-01-04
Nobody had another look at this ;-)? I just rebased it now, but it should still work unchanged and is still ready for review.