@ekpyron 2021-01-06
By the way: the numbers have slightly changed since last time I measured - apparently ``develop`` has regressed a bit further performance-wise.
I now get on develop:
```
$ time { for i in $(seq 10); do build/test/tools/isoltest -t sem*/via* --optimize &> /dev/null ; done;}

real    0m55,125s
user    0m54,665s
sys     0m0,233s
```

vs. on this branch:

```
$ time { for i in $(seq 10); do build/test/tools/isoltest -t sem*/via* --optimize &> /dev/null ; done;}

real    0m47,796s
user    0m47,389s
sys     0m0,266s
```

which is still not much, but still roughly 15% speedup.