@chriseth 2021-01-11
Haven't reviewed this in detail, but I'm not so sure about unordered_map. Can it be system-dependent? Also, about the performance comparisons: Please take into account that if you use isoltest, you are running this on many small files, which is not what we should optimize for.