@ekpyron 2021-01-11
Yeah, sure - it'd be nice to have a better set of sources to compare against... do we have that or want to come up with one?

About ``unordered_map``: do you mean whether the actual order in it is implementation-dependent or the performance? I guess for both the answer is yes, but nothing in the PR should depend on that order (I hope at least :-)).

That being said we might as well switch back to ``map`` - it doesn't make a huge difference in the last version of the PR.