@axic 2021-01-12
> Yeah, sure - it'd be nice to have a better set of sources to compare against... do we have that or want to come up with one?

The deposit contract test case was merged, that is a good one to start with.