@ekpyron 2021-01-12
> Well 15% is considerable difference

That it is, but we'll also get like 14% if we'd replace ``std::unordered_map`` with ``std::map`` and left anything else as is in this PR.