@ekpyron 2021-01-12
> > Yeah, sure - it'd be nice to have a better set of sources to compare against... do we have that or want to come up with one?
> 
> The deposit contract test case was merged, that is a good one to start with (https://github.com/ethereum/solidity/blob/develop/test/libsolidity/semanticTests/externalContracts/deposit_contract.sol)

```
$ time { for i in $(seq 100); do  buildOptimized/solc/solc --experimental-via-ir --optimize test/libsolidity/semanticTests/externalContracts/deposit_contract.sol &> /dev/null ; done ; }

real    0m30,724s
user    0m29,951s
sys     0m0,408s

$ time { for i in $(seq 100); do  build/solc/solc --experimental-via-ir --optimize test/libsolidity/semanticTests/externalContracts/deposit_contract.sol &> /dev/null ; done ; }

real    0m35,907s
user    0m35,369s
sys     0m0,425s
```

Rather consistent with the other measurements.