@axic 2021-01-12
> Rather consistent with the other measurements. But one contract doesn't really make for a good test set ;-).

Well you can pick from #10390.