@ekpyron 2021-01-12
@hrkrshnn asked me to run https://gist.github.com/hrkrshnn/24818ae3ed9f1c6f04239f1765900501, so here:
```
$ time { for in in $(seq 10); do build/solc/solc --optimize --ir-optimized DexConstraintPoly.sol &> /dev/null ; done ; }

real    1m7,018s
user    1m6,774s
sys     0m0,123s

$ time { for in in $(seq 10); do buildOptimized/solc/solc --optimize --ir-optimized DexConstraintPoly.sol &> /dev/null ; done ; }

real    1m4,477s
user    1m4,216s
sys     0m0,140s
```

That's significantly less improvement, but @hrkrshnn also already pointed out that the file doesn't have a lot of control flow.